[tufaceous] add commands to archive and extract tuf repositories (#2266)
Add two commands:

1. `tufaceous archive`, which archives a repository into a zip file.
2. `tufaceous extract`, which extracts a repository from a zip file.

## Why zip and not tar?

The main reason is that zip supports random access to files and tar does
not.

I was hoping to just be able to read the repository out of the zip file
from memory, but I ran into https://github.com/awslabs/tough/pull/563
while implementing it. Once that is resolved (or we write our own TUF
crate) it should be possible to do that.

Regardless, I don't want to foreclose that option forever, so let's just
use zip for now rather than having to deal with a migration in the
future.

Depends on #2300.