Update OPTE and use new dedicated dev node.
The kernel driver portion of OPTE (xde) no longer requires
non-upstreamed kernel bits and can run on an updated stock helios.
Thus we can remove the os-incorporation shuffling from the install
script. The uninstall script is mostly the same but tweaked a bit
to allow people with modified kernel bits to get back to stock.

Also pull in fixes for oxidecomputer/opte#335 and oxidecomputer/opte#339