Better detection of underlay devices based on system ID (#2388)
- Adds a `is_gimlet()` function that returns true on, well, Gimlets.
- Use the `is_gimlet()` to detect whether we're on a developer maching,
and thus which NICs to use in two places. The first is
`find_chelsio_links`, which has thus far been returning the emulated
Chelsios, net0 and net1. On non-Gimlets, that's still the case. On
Gimlets it returns the list of devices named `cxgbeN`, all actual
Chelsio NICs. The second is the physical link that is currently used to
determine the bootstrap address. That's net0 or cxgbe0, for non-Gimlets
and Gimlets, respecively.
- Adds a few helper functions for creating IPv6 link-locals in a
well-defined way, and factors out the name for all of them, which is now
`iface/ll`, rather than `iface/linklocal`.