[installinator] implement progress reports, client side (#2413)
Implement client-side progress reports in the installinator.

Introduce a new crate called `installinator-common`, with a `ProgressReport` type that represents a set of events. A `ProgressReport` consists of:
* completion events, which are important events that indicate completion
* progress events, which indicate general progress (for now, we only send the last event, though the data model supports multiple events)

Each event, and each report, has a timestamp associated with it (as a duration since the start of the installinator process). This helps potential users get a good sense of when things happened.

The `ProgressReport` type is much easier to work with than the types generated by progenitor (e.g. real `Duration`s), so use them in both the client and server and convert over while serializing the request in the client.