[MGS] [wicketd] Report progress from MGS to wicketd of trampoline image delivery (#2424)
During gimlet recovery, booting OS is the ultimate source of truth for
how far along the boot process is, but we can't communicate (directly)
with it while it's booting. We can get progress indirectly: this PR adds
endpoints to MGS that get and clear the most-recently-requested host
image message from the gimlet's SP. Assuming the host fetches the
trampoline image sequentially, this gives us progress-by-proxy.

We start to flesh out one of the TODOs in wicketd where we need to wait
for installinator completion: we now first wait for the first
installinator progress message (which is still a placeholder for now,
and will come in after #2413), and while we're waiting for that first
message, we poll MGS periodically to attempt to record the trampoline
boot progress.