Redesign of Wicket TUI (#2440)
This is a very large change to the look of wicket and its interaction
model. The redesign of the look is due to @benjaminleonard, although it
doesn't quite fit his figma renderings exactly :) In order to simplify
things, mouse support was removed, and the rendering model was changed
to allow using stateful widgets in the tui.rs model, where mutating
state during drawing is allowed.

The biggest change here is the fact that each `Control` no longer
maintains its location to be drawn on the screen in the form of a
`Rect`. This removes the necessity to handle resize events from the
terminal. This is in line with the tui.rs model, but eliminates the
ability to properly handle mouse events, as we cannot do rectangle
intersection on user events. It also prevents what I believe could be
somewhat simplified, and faster renderings via minimizing the amount of
computation done to that strictly necessary during rendering. For
example, if a user scrolls content in a `Paragraph` widget, if we know
they are at the bottom of the content already we don't have to redraw.
With the current model we have to redraw, because we only know the size
of the screen during the draw itself.

Oddly enough, as a result of working in these constraints and writing a
lot of this code for the second time, the way events are passed from
`Control` to inner controls and the structure of all that is much
simpler and easier to understand. I believe we could actually add the
rect/resize handling and mouse handling back with a fraction of the code
that has been deleted here. So all in all this change is a win-win,
despite my complaints about the tui.rs interaction model.

I don't think any of this is really controversial, and I"m also not too
concerned about it being looked over with a fine tooth comb. I'd vastly
prefer you run and play with it to get an idea how it works. Happy to
chat tomorrow!

---------

Co-authored-by: Benjamin Leonard <hello@benleonard.co.uk>