[Wicket] Add resize handlers to controls
This re-adds the capability for individual `Control`s to optionally
receive resize events. Such resize events allow a `Control` to compute
its layout outside of being rendered. This in turn allows the ability
to dynamically size contents based on the current state without having
to do it on every draw. Each `Control` is responsible for ensuring it's
children receive resize events with their proper `Rect` computed if
their children require this functionality.

The current code on main does not require this functionality, but it is
needed for Update. Specifically it's needed for dynamic construction  of
[`TreeItem`s](https://docs.rs/tui-tree-widget/0.11.0/tui_tree_widget/struct.TreeItem.html)
based on changing status, since we must know the width of the pane in
order to construct the tree faithfully. We could work around this by
building a fake tree structure of the same shape and then recomputing
the real tree structure on every draw, but this is likely more code
than proper resize events, and is confusing. It also costs us CPU
unnecessarily.

As a side-effect of this change, we will now be able to re-introduce
mouse events if we desire in the future.