`/v1/groups/{group}` (#2455)
We need this endpoint to power a group detail page in the console where
we fetch the group by ID and display its name and the list of users in
it.

As mentioned in a comment, groups have an `external_id` instead of a
`name` and therefore don't implement `ObjectIdentity`, which makes doing
a lookup by external ID a little less cute than lookup by name.
`external_id` is unique per silo, so looking up by it with the silo
implicit is perfectly possible in theory. _However_, unlike with `Name`,
we do not (and probably cannot) enforce that the `external_id` of a
group is _not_ a UUID, which means our usual technique of falling back
to name only once the thing fails to parse as a UUID does not work. If
an external ID was a UUID, there would be no way to fetch a group by it
because we would always assume it was one of _our_ DB UUIDs.

One option would be to add a query param that allows the caller to
specify that the given identifier is meant to be treated as the group
external ID and not our DB ID. Another option, which we can do in a
followup, is what @zephraph suggested in
https://github.com/oxidecomputer/omicron/pull/2442#discussion_r1122215806:
implement the `external_id` filter on the list endpoint instead. Because
`external_id` is unique per silo, we know this will return either 1 or 0
results. It's a bit odd but it is conceptually consistent and requires
no shenanigans.