[installinator] Rework artifact writing logic to support real hardware
We now attempt to write to both M.2 drives, if we find them. We stop
after one has succeeded and the other has either also succeeded or
gotten at least one retry chance. Retries are a little tricky because we
avoid rewriting the host image if we already succeeded on it (but failed
when writing the control plane image).