[installinator] Write host phase 2 image to both M.2s (#2474)
`installinator` now takes an `--instal-on-gimlet` flag as an alternative
to the destination path. If it's passed, we scan the hardware looking
for two M.2s (fails on non-gimlets; finding just one M.2 is okay). If we
find M.2s, after we fetch artifacts, we attempt to write the host phase
2 image directly to the boot image slice partition. For now, we don't
write the control plane (pending some decisions / info about how to
mount the zpool on the M.2s), but the logic for retrying is in place.
Assuming we find both M.2s, we will attempt to write the host and then
the control plane on each disk. We don't write the control plane until
writing the host has succeeded. We stop retrying when we've either:

1. Written both host and control plane to both M.2s successfully
2. Written both host and control plane to one M.2 successfully, and
retried the other at least once

If there is only one M.2 preset, we retry forever until it succeeds.