Add an SMF manifest for installinator, and add an `image_type` target to package-manifest (#2481)
The main intent for this PR was to add an SMF manifest to run
`installinator` as a service and to build an `installinator` package we
could include in host OS images, but it grew a bit when I went to add
`installinator` to `package-manifest.toml`. In particular, we do _not_
want installinator to be installed alongside `sled-agent` in a standard
omicron install. It should only be included in the `trampoline` image.
Therefore, this PR adds a new target key `image_type` that takes the
possible options `standard` and `trampoline`. I marked the majority of
the pre-existing packages in `package-manifest` as
`only_for_targets.image_type = standard` with the exception of
maghemite, which should be included in both standard and trampoline
images.

The end result of this is that a normal `omicron-package package` looks
almost the same, except for the additional logged key/value for the
image_type (which defaults to `standard`); note that `installinator` is
not present:

```
$ ./target/release/omicron-package package
Mar 03 21:40:40.251 DEBG target: Target({"image_type": "standard", "switch_variant": "stub"})
[00:00:04] ########################################       6/6       clickhouse: done
[00:00:08] ########################################      12/12      cockroachdb: done
[00:00:14] ########################################       1/1       crucible: done
[00:00:05] ########################################       1/1       crucible-pantry: done
[00:00:29] ######################################## 28471030/28471030 dendrite-stub: done
[00:00:00] ########################################       1/1       dns-server: done
[00:00:00] ########################################       3/3       external-dns-customizations: done
[00:00:00] ########################################       3/3       internal-dns-customizations: done
[00:00:06] ########################################       1/1       maghemite: done
[00:00:00] ########################################       4/4       omicron-gateway: done
[00:00:02] ########################################      65/65      omicron-nexus: done
[00:00:00] ########################################       5/5       omicron-sled-agent: done
[00:00:01] ########################################       4/4       oximeter-collector: done
[00:00:04] ########################################       1/1       propolis-server: done
[00:00:02] ########################################       1/1       wicket: done
[00:00:01] ########################################       4/4       wicketd: done
[00:00:01] ########################################       1/1       external-dns: done
[00:00:02] ########################################       1/1       internal-dns: done
[00:00:03] ########################################       1/1       switch-stub: done
```

Passing `--target image_type=trampoline` builds only maghemite and
installinator:

```
$ ./target/release/omicron-package --target image_type=trampoline package
[00:00:00] ########################################       3/3       installinator: done
[00:00:01] ########################################       1/1       maghemite: done
```