[wicket] Add UpdateState and poll it from wicketd
This commit builds off #2503, so that it can get access to new styles.

Update logs and artifacts are retrieved from wicketd and stored inside
the global `State` as type `RackUpdateState`.

The `FinalInstallArtifact` type represents production software installs
that the user cares about at a high level. As such it does not include
trampoline related software or delineate between phase-1 and phase-
2. Currrently,  update state is tracked in an Enum. As suggested by
@jgallagher today, this enum is likely to change to also keep track of
which specific underlying component for Host is being updated. This will
come in a follow up.

The `RackUpdateState::update_artifacts` method currently matches
`ArtifactKind` by String to `FinalInstallArtifact. This is because
`KnownArtifactKind` is not a part of the public API. This seems fine
for now, but may be worth another look. Concerns here are specifically
around the format of the strings changing, but they are not major from
my POV.

This change also makes it so that the main event channel between
`WicketdManager` and `Runner` is a tokio unbounded channel rather than a
std library unbounded channel. This allows us to get rid of translation
channels for each request inside wicketd and simplifies the code a bit.