[wicket] Add UpdateState and poll it from wicketd (#2504)
This commit builds off #2503, so that it can get access to new styles.

Update logs and artifacts are retrieved from wicketd and stored inside
the global `State` as type `RackUpdateState`.

This change also makes it so that the main event channel between
`WicketdManager` and `Runner` is a tokio unbounded channel rather than a
std library unbounded channel. This allows us to get rid of translation
channels for each request inside wicketd and simplifies the code a bit.