[tufaceous] add the ability to assemble a tufaceous archive given a manifest (#2506)
Make it easy to build out an update repository from a list of files,
specified as a manifest.

Manifests can either be:

* read in from a file (see `tufaceous/manifests/fake.toml` as an example), or
* constructed in Rust code (what permission-slip is going to be doing).

An artifact comes from a source. A source can be:

1. A file on disk.
2. A buffer in memory.
3. A fake artifact that tufaceous can hallucinate.

Most fake artifacts are generic ones without any internal structure,
with the exception of the host and control plane artifacts, both of
which have a specific format that looks like a tarball.

We haven't written the code to extract the tarballs yet; that's coming
up next.