[installinator] Configure and auto-enable smf service (#2511)
This PR makes a handful of small-ish tweaks to installinator:

* The SMF service is now set to enabled by default
* Adds all the service dependencies that sled-agent also has
* Adds a `--bootstrap-sled` flag; if passed, we act like bootstrap agent
and import+start maghemite

It also makes one related tweak to sled-agent: adding an (optional)
dependency on the `t6init` service. "Optional" here means if the service
is present, it is either running or in a "requires manual intervention
to recover" state. From sled-agent's point of view, `t6init` is the
service that causes the `cxgbeX` data links to come up, which sled-agent
expects to already exist.

Builds on #2502. Using the published
`trampoline-global-zone-packages.tar.gz` from this branch, I was able to
boot a host OS image that successfully started maghemite and
installinator.