[tufaceous] add a lib.rs next to main.rs (#2512)
We'd like to be able to run tests from other crates that run the
tufaceous binary. That isn't directly possible due to how cargo works
(you can't run a binary that's in another crate, even in integration
tests), but what we can do instead is expose tufaceous in a lib.rs and
execute commands using that.

Alternatively we could use tufaceous-lib, but I'd like to exercise as
much of tufaceous's code as possible.