[wicketd] add code to extract host and trampoline phase 1 and 2 images (#2513)
This is essentially the inverse of the code to build up the tarball,
introduced in #2506. Put this code in tufaceous-lib so that it lives in
the same place as the code that constructs fake archives, and can be
shared with other consumers.

With this, it should be possible to test that the installinator's
download functionality works.