[wicket] Start of Update Pane UI (#2507)
This builds upon #2503 and #2504. It adds the visual components to the
relevant state and wicketd interaction changes.

A tree view for update components has been added. It currently only live
updates target versions for components when a TUF repo has been
uploaded. If a TUF repo has not been added, and the user presses
`Enter`, a popup instructing the user how to upload the repo is
provided.

Currently initiating updates, and live updating progress is not
implemented in the UI, although many of the underlying mechanisms exist
already. It's expected for a follow up commit , that if `Enter` is hit
in the tree view, and the component requires updating a warning popup
will be launched. If the user selects ok, and the update starts, then
further presses of `Enter` will toggle log views of the ongoing update.
If the update has failed, we may also want a way to show this. I need to
think a bit about these interactions and likely play with them once
implemented.

Actual known versions of running software from inventory also needs to
be put into this view. This information for SP and RoT is currently
known from MGS, but we will soon have additional info available from
sled-agents. It's unknown if we will actually plumb this through
immediately.

The system version for an update is currently listed as "UNKNOWN" in the
lower right of the screen (in the status bar). This is because we don't
currently have that info inside the TUF repo. This will be fixed in a
follow up as well.

Additional styling to fade either the sidebar for selecting panes, or
the panes themselves was added to make it more apparent the context the
user is operating in. It also makes the rack rendering look cool :)