[wicket] switch to being configured via sshd ForceCommand (#2520)
Previously, wicket was expecting to be run as a login shell. However, we
decided in
https://github.com/oxidecomputer/minimum-upgradable-product/issues/20
that we'd instead configure wicket to be run via sshd `ForceCommand`.

The mechanism for doing that is slightly different, using
`SSH_ORIGINAL_COMMAND` rather than `-c`. Support this new mechanism and
remove support for the old login shell mechanism.

Tested through the instructions included in the PR.