[wicketd] Add installinator completion events to update event log (#2519)
The update event log types are getting a little hard to follow and are
probably ripe for a rework. This PR renames a few types and enum
variants:

* `UpdateEventSuccessKind` -> `UpdateNormalEventKind`, which also gained
an `::InstallinatorEvent` variant. This rename is because installinator
events can represent transient failures that will be retried, which
didn't seem to fit under either `Success` or `Failure`
* `UpdateEventKind::Success` -> `UpdateEventKind::NormalEvent` to go
with the previous rename
* `UpdateEventKind::Failure` -> `UpdateEventKind::TerminalFailure` to
clarify that these result in the update process stopping until manual
intervention is applied to restart it / debug it

A couple of minor bumps in the road:

* Reporting age of events from installinator through wicketd is a little
weird (see the bit in the diff where we convert to an `Instant`)
* Installinator's `CompletionEventKind` has variants that cover both
success and failure cases - maybe the wicketd update log should use a
flatter representation like this too?
* wicketd's openapi is now effectively "re-exporting"
`CompletionEventKind` from installinator. That seems... okay? but a
little weird.