[common] add a system version to ArtifactsDocument (#2523)
Since artifacts.json is going to represent a single rack update, we need
to be able to specify a system version associated with it. Add a
system_version field which is of type SemverVersion.

Also update tufaceous init to take the system version as an argument,
and tufaceous assemble to use the system version specified in the
manifest.