[mgs] Return detailed error messages with HTTP 503 codes
This is primarily for plumbing through to wicketd -> wicket, but will
probably be useful to nexus in the future. MGS returns HTTP 500-level
codes (mostly 503 unavailable) when it has issues communicating with an
SP; previously, these errors did not include any details in the response
body. We now include an error message string.