[package] Add build targets notion into 'omicron-package' (#2529)
Previously, `omicron-package` used a `--targets` flag that acted as a
very weakly-typed key-value set.

This worked, but had a few problems:

- Being weakly typed, it was easy to get one of the characters wrong.
This could result in a silent error, rather than flagging the issue.
Note: this issue also applies to changing the value of any of the
key-value pairs.
- For backward compatibility, all the keys were *optional*. For values
like "is this a gimlet" vs "is this not a gimlet", we can pick defaults,
but it feels a little hidden to never have the user opt-in to the target
being built.
- It involved a lot of typing, and required repeating the "target" value
at each invocation of `omicron-package`. Since the target value changes
many commands, not *just* packaging, it can be irritating to re-supply
the full list of key/value pairs.

## TL;DR

- Run `cargo run --release -p omicron-package -- -t default target
create` once, and a `default` target will be created for you. Add the
`-h` argument to that `target create` command if you want to see which
defaults you picked.
- The `-t <TARGET>` command can be used to other `omicron-package`
subcommands, and it'll use the configuration you picked. If you omit it,
it'll use whatever last target value you created or set explicitly.
- If you want to re-visit your target choice, the `omicron-package
target` subcommand lets you view / modify your selection.

## The longer version

In this PR, we introduce the notion of "build targets" explicitly.

Rather than supplying `--target <LIST OF KV PAIRS>`, the target flag is
used as a label. Additionally, when parsing targets, we use strong
types, which prevents errors, and future-proofs the addition of new
build target types.

### The `target` subcommand

The `target` subcommand allows adding, viewing, and deleting targets:

```
Usage: omicron-package target <COMMAND>

Commands:
  create  Creates a new build target, and sets it as "active"
  list    List all existing targets
  set     Activates a build target, with a symlink named "active"
  delete  Delete an existing target
  help    Print this message or the help of the given subcommand(s)
```

For example, to create a new image, intended for utilization on an
in-rack gimlet:
```
$ cargo run --release -p omicron-package -- -t deploy target create -i standard -m gimlet -s asic
Created new build target 'deploy' and set it as active
```

As another example, to create a recovery image:
```
 $ cargo run --release -p omicron-package -- -t recovery target create -i trampoline -m gimlet -s asic
Created new build target 'recovery' and set it as active
```

Information about the created targets can be displayed:

```
$ cargo run --release -p omicron-package -- target list
             deploy: image=standard machine=gimlet switch=asic 
SELECTED >>> recovery: image=trampoline machine=gimlet switch=asic
```

The "SELECTED" target is tracked via a symlink in `out/target/active`,
and can be changed with:
```
$ omicron-package -t <TARGET> target set
```

### Using targets

All commands which previously operated on a list of keys can now simply
reference an existing target name:

```
# Uses the 'recovery' target we made earlier
$ cargo run --release -p omicron-package -- -t recovery package
...
```

As a convenience, the target can be omitted if it has been selected:
```
$ cargo run --release -p omicron-package -- -t recovery target set
Set build target 'recovery' as active
$ cargo run --release -p omicron-package -- package
...
```