wip - start recording history of events so we can replay them and visually debug.
This is wip because of the following:

For checkpoints to work properly, all state including visual state must
live in the global `State`. We can store this mapped by a re-introduce
ControlId if we want and expose it to controls that way in their
callbacks, or we could do the simple thing and just add all relevant
state from Controls and panes to the global `State` directly.