Delete organizations (#2626)
Per RFD 332. Goodbye. Very little interesting going on, just a lot of
little things. I'll see if there's anything to comment on.

One mildly interesting takeaway I have is that almost nothing in our
integration tests was depending on the existence of orgs in a meaningful
way, i.e., besides just using them to specify a project. This is a good
confirmation that they don't need to exist.