[wicket] Introduce wicket-dbg
This builds on top of #2638

This commit introduces the wicket-dbg package which contains two new
binaries `wicket-dbg-server` and `wicket-dbg`. `wicket-dbg` is an
interactive CLI for controlling the `wicket-dbg-server`, which is
capable of loading and playing recordings from previous wicket dumps
as of this commit. `wicket-dbg` also allow single stepping, jumping,
changing the rate of playback, and retrieval of both wicket state and
debug server state.

One major limitation is that resize currently does not work. In
particular if your terminal running the `wicket-dbg-server` is smaller
than the terminal of the recording, the server may crash. Additionally,
it's unknown what the effect of resizes during recoring will be during
playback. This is somewhat difficult to fix, and not immediately
relevant since we anticipate using this debugger mostly during iterative
development right now. Note however, that the debugger, modulo sizing
is currently capable of playing recordings dumped from production wicket
systems, and so we will likely want to fix this for that use case.