[wicket] Start recording event history (#2636)
We are starting to record events in memory so that we can dump a
snapshot and replay it for visual debugging.

The events must replay determinisitically, and so events that contained
`Instants` were changed to use `Durations` resulting in some
corresponding change elsewhere. This actually fixed a bug I noticed
periodically in display of the countdown when MGS was offline for a
while, and countdown would move at a weird rate.

Right now the recording isn't 100% deterministic because there is some
state stored within the UI panes. We should move this state into the
global `State` to make it deterministic. This actually doesn't matter
for short recordings, as we always start from newly initialized data.
But if we ever have enough data to exceed our event storage, we snapshot
`State`, which doesn't capture what's in the panes and may trigger
anomalies.